% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{constrain}
\alias{constrain}
\title{Constrain}
\usage{
constrain(x, minimum, maximum)
}
\arguments{
\item{x}{dataset that is to be constrained}

\item{minimum}{Value of the lowest acceptable number, before being replaced by this lower bound}

\item{maximum}{Value of the highest acceptable number, before being replaced by this upper bound}
}
\description{
constrains a dataset
}
\keyword{internal}
