% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bin.R
\name{convert.hexstream}
\alias{convert.hexstream}
\title{convert.hexstream}
\usage{
convert.hexstream(stream)
}
\arguments{
\item{stream}{Data from GENEActiv .bin file feed into this stream}
}
\value{
Returns a decrypted raw data
}
\description{
internal function for read.bin
}
\keyword{internal}
