% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.time.R
\name{convert.time}
\alias{convert.time}
\title{convert.time}
\usage{
convert.time(x, format = NULL)
}
\arguments{
\item{x}{Object to process. For convert.time, must be numeric.}

\item{format}{A character string indicating the form of output. See \code{\link[base]{strptime}} for details. If
NULL, will be automatically chosen.}
}
\description{
extends time display from package chron to use h:m:s for > 1 day times.
}
\details{
convert.time converts numerics to GRtime objects. The format argument allows a format string to be attached
specifying the default format to display in. as.GRtime is a wrapper to convert.time, that when supplied with
character input, coerces the value first to numeric using parse.time.
}
