% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(X, length = (length(X) * 100))
}
\arguments{
\item{X}{object to expand}

\item{length}{length to expand object}
}
\description{
expand an object X
}
\keyword{internal}
