% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENEActiv.calibrate.r
\name{g.downsample}
\alias{g.downsample}
\title{g.downsample}
\usage{
g.downsample(sig, fs, ws3, ws2)
}
\arguments{
\item{sig}{Vector of numeric values}

\item{fs}{Sample frequency}

\item{ws3}{epoch size, e.g 5 seconds}

\item{ws2}{epoch size, e.g 90 seconds}
}
\description{
Downsamples a vector of numeric values at three time resolutions:
1 seconds, ws3 seconds, and ws2 second. Function is not intended for direct interaction
by package end user
}
\details{
List with three object: var1, var2, and var3 corresponding to downsample
time series at 1 seconds, ws2 seconds, and ws3 seconds resoluton, respectively
}
\author{
Vincent T van Hees <vincentvanhees@gmail.com>
}
\keyword{internal}
