% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stftmod.R
\name{getfreqs}
\alias{getfreqs}
\title{getfreqs}
\usage{
getfreqs(x, frequencies)
}
\arguments{
\item{x}{stft object}

\item{frequencies}{frequencies range}
}
\description{
gets fft components corresponding to frequencies
}
\keyword{internal}
