% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%bq\%}
\alias{\%bq\%}
\title{\%bq\%}
\usage{
X \%bq\% y
}
\arguments{
\item{y}{Second object to pass}

\item{x}{First object to pass}
}
\description{
quantile version of bt
}
\keyword{internal}
