% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%bt\%}
\alias{\%bt\%}
\title{\%bt\%}
\usage{
X \%bt\% y
}
\arguments{
\item{y}{Second object to pass}

\item{x}{First object to pass}
}
\description{
between' operator for convenience
takes \code{[min, max)}, or \code{c("[", min, max, "]")} style second terms
default is \code{[min, max]} for c(,) terms
}
\keyword{internal}
