% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stftmod.R
\name{hanning.window}
\alias{hanning.window}
\title{Hanning Window}
\usage{
hanning.window(n)
}
\arguments{
\item{n}{number of points inside the window}
}
\description{
A hanning window used by the STFT function
}
\keyword{Internal}
