% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bin.R
\name{is.POSIXct}
\alias{is.POSIXct}
\title{Utility functions to be used within GENEAread}
\usage{
is.POSIXct(x)
}
\arguments{
\item{x}{Time object passed to check class}
}
\description{
To check the timestamps are of the correct format when using within read.bin
}
\details{
utility function for checking timestamps
}
\keyword{internal}
