% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pseudolines}
\alias{pseudolines}
\title{pseudoline}
\usage{
pseudolines(x, y = NULL, max.shift, new = FALSE, ...)
}
\arguments{
\item{x}{sorted vector}

\item{y}{response variable}

\item{max.shift}{maximum to shift data}

\item{new}{Create new plot}
}
\description{
Plot a line graph, with breaks when things change too much
(assume x is sorted)
}
\keyword{internal}
