% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{removeZero}
\alias{removeZero}
\title{removezero}
\usage{
removeZero(obj)
}
\arguments{
\item{obj}{object to remove zeros from}
}
\description{
removes zero's from an object
}
\keyword{internal}
