% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seq.log}
\alias{seq.log}
\title{seq.log}
\usage{
\method{seq}{log}(from = 1, to = 1, length.out = 50,
  add.zero = FALSE, shifting = 0)
}
\arguments{
\item{from}{Value to Create the sequence from. default to 1.}

\item{to}{Value to create the sequence to. default to 1}

\item{length.out}{length of list outputted}

\item{add.zero}{include 0 at the start}

\item{shifting}{Shifting the sequence by this value}
}
\description{
Create a sequence of log values
}
\keyword{internal}
