% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{shift}
\alias{shift}
\title{shift}
\usage{
shift(x, offset, expand = 0, fill = c("zero", "edge", "loop", "mean"))
}
\arguments{
\item{x}{data object to process}

\item{offset}{2 integers to offset a dataset}

\item{expand}{2 integer list to expand the dataset}

\item{fill}{method to run on the shift}
}
\description{
shift an object x in a direction/expansion
}
\keyword{internal}
