% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stftmod.R
\name{subs}
\alias{subs}
\title{subs}
\usage{
subs(x, a, b)
}
\arguments{
\item{x}{object to subset}

\item{a}{Start point to separate the object}

\item{b}{End point to separate the object}
}
\description{
subsets a proportion of the dataset, or a certain length of the dataset starting at a
specific proportion position
}
\keyword{internal}
