% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{svm}
\alias{svm}
\title{sum vector magnitude}
\usage{
svm(obj, sqrt)
}
\arguments{
\item{obj}{AccData object}

\item{sqrt}{Function to use to calculate SVM}
}
\description{
svm acts identically to 'mean.epoch', with the epoch set to the sampling period. In other words, it computes the
instantaneous sum of vector magnitudes of the acceleration at each record point. The function takes "AccData",
array and vector input. Note that if provided with an array with 4 or more columns, columns 2 to 4 are used -- the
first column is regard as a timestamp and hence ignored.
}
\examples{
dat <- read.bin(system.file("binfile/TESTfile.bin", package = "GENEAread")[1], calibrate = TRUE)
svm(dat)

}
