% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stftmod.R
\name{uniform.window}
\alias{uniform.window}
\title{Uniform Window}
\usage{
uniform.window(n)
}
\arguments{
\item{n}{number of points inside the window}
}
\description{
A uniform window used by the STFT function
}
\keyword{Internal}
